+++ src/daemon/touch-manager.cpp
@@ -26,7 +26,7 @@ namespace iptsd::daemon {
 
 TouchManager::TouchManager(Config conf)
 	: size(), conf(conf), inputs(IPTS_MAX_CONTACTS), last(IPTS_MAX_CONTACTS),
-	  distances(IPTS_MAX_CONTACTS * IPTS_MAX_CONTACTS)
+	  distances(static_cast<std::size_t>(IPTS_MAX_CONTACTS) * IPTS_MAX_CONTACTS)
 {
 	for (i32 i = 0; i < IPTS_MAX_CONTACTS; i++) {
 		this->last[i].index = i;
